requirements.txt
# API Dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
slowapi==0.1.7
starlette==0.27.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.24.1
bcrypt==4.0.1
PyJWT==2.8.0

# MongoDB
pymongo==4.5.0

# Vector Database
chromadb==0.4.15
langchain-community==0.0.13
langchain-chroma==0.1.4
langchain-huggingface==0.0.2
sentence-transformers==2.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# PDF Processing
pdfminer.six==20221105
PyPDF2==3.0.1

# AI/ML Dependencies
torch==2.0.1
transformers==4.33.2
accelerate==0.23.0
google-generativeai==0.3.1

# UI Dependencies
streamlit==1.27.0
pandas==2.1.0
plotly==5.17.0
numpy==1.25.2
matplotlib==3.8.0
seaborn==0.12.2
altair==5.1.1
pillow==10.0.1

# Environment
python-dotenv==1.0.0

# NLTK for text processing
nltk==3.8.1